Implement the snapping behaviors for all direction
Pressing "[modifier keys] + UpArrow + LeftArrow" will now snap the window to the top-left quadrant of the screen. It also works for top-right, bottom-left and bottom-right quadrants of the screen. Also, pressing "[modifier keys] + UpArrow" twice will but the window in windowed fullscreen mode (i.e. the dock and the menu bar are still there).