Implement the desktop background when showing the suggested snaps window
By masking the opened windows with a pixel-perfect screenshot of the desktop, it gives the user the illusion that we are really showing him the desktop. In addition to the black overlay, this feature gives the user a good idea of exactly where the suggested snap window would be positioned after selecting it. Finally, the fadein/fade-out adds a nice touch to make the user experience more enjoyable.